Add Repository Labels (#2)
# Pull Request

## Description

This pull request introduces a new configuration for GitHub labels and a
workflow to synchronize these labels automatically. The main changes
include adding a set of default labels and creating a workflow to keep
these labels in sync with the configuration file.

Configuration of GitHub labels:

*
[`.github/other-configurations/labels.yml`](diffhunk://#diff-c1b32711b0f3da3f5c7007cfaa4d9e94ae5430fea1f4e1256f6210bc6988553dR1-R63):
Added default labels for issues and pull requests, categorized by type,
language, and size.

Workflow for label synchronization:

*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6R1-R27):
Created a new GitHub Actions workflow to synchronize labels whenever
changes are pushed to the main branch.