Add Labels to Pull Requests (#3)
# Pull Request

## Description

This pull request introduces automation for labeling pull requests based
on file changes and size. The main changes include the configuration of
auto-assigned labels and the setup of a GitHub Actions workflow to
manage these tasks.

Configuration of auto-assigned labels:

*
[`.github/other-configurations/labeller.yml`](diffhunk://#diff-b490f987d9b055ce6e06e37d13763d61f6e0d5ee349b390bbb910f8d88ff36b9R1-R44):
Added a configuration file to automatically assign labels to pull
requests based on the types of files changed. This includes labels for
documentation, markdown, dependencies, shell scripts, GitHub Actions,
end-to-end tests, dashboard changes, and git hooks.

Setup of GitHub Actions workflow:

*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR1-R34):
Created a workflow to label pull requests automatically. This workflow
triggers on pull request events (opened, edited, synchronized) and
includes steps to apply labels based on the configuration file and to
add size labels based on the number of lines changed.