Add Dependabot Configuration (#5)
# Pull Request

## Description

This pull request introduces a new configuration for Dependabot to
automate dependency updates for GitHub Actions. The main change involves
adding a new configuration file to specify the update schedule and
parameters.

Configuration for automated dependency updates:

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28R1-R21):
Added a new configuration file to automate dependency updates for GitHub
Actions with a weekly schedule, targeting the main branch, and grouping
updates for patch and minor versions.