Run Project Status Checker (#7)
# Pull Request

## Description

This pull request introduces a new GitHub Actions workflow to automate
the project status checking process. The new workflow is configured to
run on a schedule, on demand, and on pushes to the main branch.

New GitHub Actions workflow:

*
[`.github/workflows/run-project-status-checker.yml`](diffhunk://#diff-acd8df7c3dec8258b0608875cd6decf061ccb01c964b4da34cc29df6257bda67R1-R25):
Added a new workflow named "Run Project Status Checker" that runs on a
schedule (every hour), on workflow dispatch, and on pushes to the main
branch. This workflow checks out the repository and uses the
`JackPlowman/project-status-checker@v1.0.0` action.