Add Code of Conduct (#15)
# Pull Request

## Description

This pull request introduces a new Code of Conduct document to the
project. The document outlines the community's standards and
expectations for behavior, as well as the enforcement responsibilities
and guidelines for addressing violations.

Key changes include:

*
[`docs/CODE_OF_CONDUCT.md`](diffhunk://#diff-5f540ea627655d624f2992755195b9cda3bb919cad13bc3e002cf2bdc520beaeR1-R131):
Added a comprehensive Contributor Covenant Code of Conduct, including
sections on Our Pledge, Our Standards, Enforcement Responsibilities,
Scope, Enforcement, and Enforcement Guidelines.