Add CodeQL Checks (#26)
# Pull Request

## Description

This pull request includes updates to the CodeQL workflow in the
`.github/workflows/code-checks.yml` file. The changes involve updating
the version of the `upload-sarif` action and adding a new job for CodeQL
analysis.

Updates to CodeQL workflow:

* Updated the `upload-sarif` action to version `v3.28.9` from `v3.27.9`.
* Added a new job named `run-codeql-analysis` to perform CodeQL
analysis, including steps to checkout the repository, initialize CodeQL,
and perform the analysis.