Update Dependency Review (#33)
# Pull Request

## Description

This pull request adds a new job to the GitHub Actions workflow to
perform a dependency review. The most important change is the inclusion
of a `dependency-review` job in the
`.github/workflows/pull-request-tasks.yml` file.

Changes to GitHub Actions workflow:

*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR36-R51):
Added a new `dependency-review` job that checks out the repository and
runs the `actions/dependency-review-action` to review dependencies and
comment on the pull request if any issues are found.