Add Git Hooks (#34)
# Pull Request

## Description

This pull request introduces several new utilities and improvements to
the development workflow by adding commands for code formatting, code
style checking, and git hooks. The most important changes include the
addition of Prettier and Justfile commands for formatting and checking
code, as well as the creation of pre-commit and post-commit git hooks.

Development workflow improvements:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR1-R36):
Added commands for Prettier to check and format code, commands for
Justfile formatting, and a command to run gitleaks for detecting
secrets.
*
[`githooks/pre-commit`](diffhunk://#diff-d5fe9d324948511c959e14d0aa311bb057ce7236626bb86fa51351dc26845a4dR1-R30):
Created a pre-commit hook to run checks for gitleaks, Prettier, and
Justfile formatting, ensuring code quality before commits are made.
*
[`githooks/post-commit`](diffhunk://#diff-434d9a5655c19f7989a916bcec6e2e60d68a96eae07a53044290623181b520e7R1-R4):
Created a post-commit hook to update the git index after each commit.