Check Justfile Format (#35)
# Pull Request

## Description

This pull request includes a new job to check the format of the
`Justfile` in the GitHub Actions workflow configuration. The most
important changes are:

New job addition:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R37-R50):
Added a new job named `check-justfile-format` to the GitHub Actions
workflow. This job sets up the `Just` tool and runs a format check on
the `Justfile`.