Update SuperLinter EditorConfig Checker (#45)
# Pull Request

## Description

This pull request includes a couple of changes to configuration files
for code checks. The most important changes are the renaming of a
configuration file and the addition of a new environment variable to the
GitHub Actions workflow.

Configuration file changes:

*
[`.github/super-linter-configurations/.editorconfig-checker.json`](diffhunk://#diff-926aaab00380cdd79b8c5069e4eb6d6a7b0a730b2cfd8254e271be7844a0bd10L5-R5):
Renamed from `.ecrc` and corrected a typo in the `SpacesAfterTabs` key.

GitHub Actions workflow changes:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R36):
Added `EDITORCONFIG_FILE_NAME` environment variable to the `jobs`
section.