Add GitHub Action Step Name (#58)
# Pull Request

## Description

This pull request includes a minor update to the
`.github/workflows/sync-labels.yml` file to enhance the clarity of the
GitHub Actions workflow by adding a descriptive name to a step.

Workflow improvements:

*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6L25-R26):
Added the name "Sync labels" to the GitHub Actions step that uses the
`micnncim/action-label-syncer` action, improving readability and
maintainability of the workflow.