Add Pinact Verify Job (#61)
# Pull Request

## Description

This pull request adds a new job to the GitHub Actions workflow for
verifying Pinact. The most important change is the introduction of the
`pinact-verify` job, which includes steps for checking out the
repository and installing the latest version of `uv`.

### Workflow updates:
* `.github/workflows/code-checks.yml`: Added a new job named
`pinact-verify` to the workflow. This job runs on `ubuntu-latest` and
includes steps to:
  - Checkout the repository using `actions/checkout@v4.2.2`.
- Install the latest version of `uv` using
`astral-sh/setup-uv@d4b2f3b6ecc6e67c4457f6d3e41ec42d3d0fcb86`.