Add actionlint Git Hook (#130)
# Pull Request

## Description

This pull request introduces a new `actionlint` check to the development
workflow to ensure GitHub Actions workflows are validated for best
practices. The changes include adding a new `actionlint-check` command
to the `Justfile` and integrating it into the `lefthook` pre-commit
hooks.

### Additions to the development workflow:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR49-R56):
Added a new `actionlint-check` command to run `actionlint` for
validating GitHub Actions workflows.
*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dR23-R24):
Integrated the `actionlint-check` command into the `pre-commit` hooks to
ensure it runs automatically during pre-commit checks.