Add GitHub Workflow Permission Notes (#169)
# Pull Request

## Description

This pull request updates several GitHub Actions workflow files to
clarify the purpose of permission settings by adding explanatory
comments. These comments help future maintainers understand why each
permission is needed for specific jobs, improving overall workflow
readability and maintainability.

Permissions documentation improvements:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L23-R24):
Added comments to the permissions for `actions`, `pull-requests`, and
`security-events` in the "Common Code Checks" job, and for
`security-events` in the "CodeQL Analysis" job to clarify their usage.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L23-R24)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L34-R33)
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfL13-R13):
Added a comment to the `pull-requests: write` permission in the "Common
Pull Request Tasks" job explaining it is for writing labels and comments
on PRs.
*
[`.github/workflows/run-project-status-checker.yml`](diffhunk://#diff-acd8df7c3dec8258b0608875cd6decf061ccb01c964b4da34cc29df6257bda67L17-R17):
Added a comment to the `contents: write` permission in the "Run Project
Status Checker" job to explain it is for committing changes to the
repository.
*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6L18-R18):
Added a comment to the `pull-requests: write` permission in the
"Configure Labels" job to clarify it is for writing labels to the
repository.

Other minor workflow cleanup:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L13):
Removed the unused `packages: read` permission from the top-level
permissions section.